+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added [ysort], [ycalc], [fval] functions.
 - Added [wrap] function.
-- Added [percentile] supports multiple value fields.
+- Added [percentile] supports multiple value fields. Missing/Empty value detection/removal/replacement.
 
 
 ## [2.0.0]